* commit '0f3d5387524f6ced26b1792a2731fdfae0701d0f':
  - Optimal Code
  fix mini check
  - Fix mini UI
  - Fix phoneNumber with +84 - Optimal Code
  - Fix text flow request
  - Optimal Code
  - Optimal Code - Change type data salary

+++ FinPlus/Model/Browwer/Info/BrowwerJobInfo.swift
@@ -31,7 +31,7 @@ public struct BrowwerJobInfo {
   // MARK: Properties
   public var position: Int?
   public var companyPhoneNumber: String?
-  public var salary: Int?
+  public var salary: Double?
   public var company: String?
   public var jobType: Int?
   public var jobAddress: BrowwerAddress?
@@ -59,7 +59,7 @@ public struct BrowwerJobInfo {
   public init(json: JSON) {
     position = json[SerializationKeys.position].int
     companyPhoneNumber = json[SerializationKeys.companyPhoneNumber].string
-    salary = json[SerializationKeys.salary].int
+    salary = json[SerializationKeys.salary].double
     company = json[SerializationKeys.company].string
     jobType = json[SerializationKeys.jobType].int
     jobAddress = BrowwerAddress(json: json[SerializationKeys.jobAddress])
